PVW-4721 + PVW-4626: Render array values and empty string
* Do not show item count for single-item arrays in DataAttributeRow

The DataAttributeRow widget will now only display the item count for ArrayValue attributes if the array contains more than one item. Goldens were also updated.

* Add Golden tests for DataAttributeRow

This commit adds a suite of golden tests for the `DataAttributeRow` widget. These tests cover various attribute value types, including:

- String (including empty string)
- Boolean (true and false)
- Number
- Date
- Null
- Array (empty and multi-value)

* Improve rendering of array attributes in DataAttributeRow

This commit refactors the `DataAttributeRow` widget to improve the display of attributes with array values.

Previously, array attributes were simply displayed with a count of their items. Now:
- Each item in an array is displayed on a new line, prefixed with a bullet point.

To achieve this, the following changes were made:
- Introduced a `BulletListDot` widget for consistent bullet point rendering.
- Updated `DataAttributeRow` to recursively build subtitles for array items.
- The `MissingAttributesSheet` now also utilizes the new `BulletListDot` widget.

* Display attribute count in label for ArrayValue

For attributes of type `ArrayValue`, the `DataAttributeRow` will now display the number of items in the array within parentheses next to the attribute label.

* Style empty string attribute values the same as null values

When an attribute value is an empty string, it should be displayed with the same text style as a null value. This change ensures consistency in how empty or missing data is presented.

* Use locale extension to get l10n and format ArrayValue

This commit introduces the `l10n` getter on `LocaleExtension` to simplify accessing localized strings.
It also updates `AttributeValueFormatter` to correctly format `ArrayValue` attributes by recursively calling `formatWithLocale` for each element in the array. Previously, it would directly convert array elements to strings.

* Format empty string values as "Left empty"

This commit introduces a new localization string `cardValueEmpty` and updates the `AttributeValueFormatter` to use it when a `StringValue` contains an empty or whitespace-only string.

Issue: PVW-4626

* Move array display logic to AttributeValueFormatter

The toString() is not intended for display to the user, as such move the display
logic into the value formatter.

* Refactor: Simplify AttributeValueFormatter

Reduce code duplication by redirecting to the same formatWithLocalse() method for
both entry points.

* Add an ArrayValue to the mock data