PVW-4695: App bar / Fade in tweaks
* Update _buildTitle() of DisclosureScreen

Use context.watch() for better readability and make sure we exhaustively match on
the DisclosureState to render all the title options. Also update teh fade in offset
for the check organization page so it fades in correctly with the extended (logo + title)
header.

* Refactor IssuanceScreen AppBar title fading

This commit changes the title fading behavior in the `IssuanceScreen`.

Previously, the `WalletAppBar`'s default `fadeInTitleOnScroll` behavior was used.
Now, `fadeInTitleOnScroll` is disabled, and the `_buildTitle` method wraps the `TitleText` in a `FadeInAtOffset` widget.

This allows for custom fade-in offsets based on the current `IssuanceState`. Specifically, the `IssuanceCheckOrganization` state, which has a taller header with a logo, now has a later fade-in offset for its title.

Additionally, the `ScrollOffset` is now reset between pages in the `BlocConsumer` listener to ensure correct fading behavior when navigating between issuance steps.

* Only show fade-in title when scroll area is large enough

* Wrap IssuanceScreen with ScrollOffsetProvider

This change is necessary to ensure the WalletAppBar's fade-in title animation works correctly within the IssuanceScreen.