PVW-4689: Third-party developer onboarding
* Spelling fixes

* Erronuous wow

* Forgotten changes (save before commit)

* Review comments processed

* More generic architecture and development sections

* Fixed typo

* Added TODO notes regarding issuer and verifier development

* First verified working onboarding instructions

* Onboarding.md work

* About page in english

* Added about body

* Various link fixes

* Moved documentation to wallet_docs

* Use .venv instead of venv

* First outline of onboarding.md, preamble

* Fixed various links, split out sphinx doc howto

* Fixed various links to docs
* Made a static link relative
* Moved sphinx documentation
* Dedicated about page

* New wallet-logo location

* Re-layout of documentation sections

This commit creates a sectioned toctree which takes into consideration
that we now document the development of both a verifier *and* an issuer.

In addition, we have added a lot of documentation with regards to what
kind of specifications we use, with regards to generic issuance and
disclosure-based issuance, and finally, a new process for onboarding.

We now distinguish the following sections:

  * about
  * get-started
  * architecture-diagrams
  * development-notes
  * way-of-working
  * release-notes

* Added sphinx-autobuild

Added sphinx-autobuild and updated README.md to show how to use it.
Additionally, formatted and slightly improved README.md.

* Replaced PNG logo with SVG variant, no __pycache__

Replaced wallet.png with wallet.svg, added __pycache__ to .gitignore.
The SVG variant is scaled a little larger but it looks good (the PNG
was 128x128, the SVG is of course infinite, but configured as 512x512).

* Get and set documentation version with version.sh

Since the excellent work to get our documentation publishable by using
the Sphinx framework was done, we also have a synchronizable version in
the documentation/conf.py which was not handled by scripts/version.sh.

This commit:

  * conf.py: Normalizes version and release
  * conf.py: Normalizes author, project and copyright
  * version.sh: Add getter for documentation version
  * version.sh: Add setter for documentation version