Implement DCQL disclosure response validation
This implements matching of a disclosure response to a of DCQL
credential queries, which replaces `DeviceResponse` validation in the
`mdoc` crate. This also inverts the relationship between the `dcql` and
`mdoc` crates, so that the former has the latter as a dependency. In
order to archieve this, some test code had to be moved to the `dcql`
crate.

Note that `openid4vc` currently constructs a response by repeating the
credential query ids from the request. Once `openid4vc` uses DCQL
instead of Presentation Exchange this will no longer be necessary.