PVW-4824: Update stepper count and remove from error pages
* Fix lint issue

* Add tests for SetupHelper

- Provide a sane default for tests
- Verify the SetupHelper behaves as expected

* Refactor onboarding stepper to be dynamic

The total number of steps in the onboarding flow can now change based on whether biometrics are available on the device.

* Refactor: Adjust setup stepper

- Remove stepper from loading states to align with Figma
- Avoid dedicated step for biometrics to align with current Figma,
this could eventually be improved my making the total steps dynamic based on
biometric availability and then adding a dedicated biometrics step.

* Refactor: Enforce `FlowProgress` to be 1-based and update tests

* Refactor: Make `stepperProgress` nullable in `WalletPersonalizeState`

Remove the stepper from the possible error states during the wallet_personalize flow.

* Refactor: Update `_buildFlowProgress` in `ChangePinScreen`

Return null for error states and return directly from the switch statement so we can provide
const FlowProgress objects.

* Refactor: Make stepper nullable in disclosure flow

Update the DisclosureState to provide 'null' as a value for the stepper in the disclosure flow,
this causes the stepper to be invisible in error states, as it only led to confusion there.

* Refactor: Make stepper nullable in issuance flow

Update the IssuanceState to provide 'null' as a value for the stepper in the issuance flow,
this causes the stepper to be invisible in error states, as it only led to confusion there.

* Refactor: Adjust step numbering in Renew PID flow

- Update the total number of steps, so that we can use 1-based indexing, which is more intuitive for the
end user (since the steps are announced for a11y).
- Remove stepper from error pages: Error pages now no longer feature the stepper, since the step x out y
does not make sense in this context and only led to confusion.