PVW-4836: Introduce `FocusOnInit` widget
* Update goldens after merging main

* Provide Key to BackIconButton

Update the definitions of BackIconButton to provide a key based on the page's
state. This way the internal focus logic can trigger accordingly on page transitions
because it's treated as a new BackButton (and thus re-requests focus).

* Refactor IntroductionScreen for initial focus

- Provide the title and description as a widget for more flexibility.
- On the first page, request focus on the page's title so the app is greeted
by the title when initially opening the app.

* Prevent call to `sendSemanticsEvent` when LoadingPage is no longer mounted

* Refactor focus behaviour for BackIconButton

Make it so that the BackIconButton by default attempts to gain a11y focus,
this helps reset the focus to the top of the page on page transitions. To make
sure behaviour works as expected a dedicated Key is provided where needed, to make
sure the internal FocusOnInit widget is triggered to regain focus accordingly.

* Disable autofocus on PIN keyboard when screen reader is enabled

This change disables the autofocus behavior of the PIN keyboard when a screen reader is active. This prevents the screen reader from immediately focusing on the keyboard, which could interrupt the announcement of other important content on the screen.

* Wrapped government logo with `ExcludeSemantics`

The government logo intervened with the accessibility navigator ordering, opting to ignore
it for a more accessible experience, seeing as this logo is mostly decorative and we will gain
a dedicated 'trust mark' page in the future.

* Add `requestFocus` flag to `FocusOnInit`

* Focus on Pin title during setup security

* Add `FocusOnInit` widget to focus on a widget when it is first built

This widget can be used to wrap another widget. When the widget is first built, it will be focused after a short delay. This is useful for accessibility, as it allows screen readers to announce the new content.