PVW-4850: Improve Menu a11y
* Conditionally exclude AppImage from semantics

When `altText` is null, the `AppImage` widget will be excluded from the semantics tree. This prevents accessibility services from announcing images that are purely decorative or don't have a meaningful description.

* Refactor MenuScreen WCAG announcement

- Remove `AnnounceOnInitMixin` and use `DoOnInit` widget instead to announce the number of menu items when the screen is initialized.
- This change improves the accessibility of the MenuScreen by providing context to users relying on screen readers.

* Refactor SettingsScreen to use ListView

This commit refactors the `SettingsScreen` to use a `ListView` instead of a `CustomScrollView` with a `SliverList`. This simplifies the widget structure for displaying the settings content and improves a11y
navigation.

* Introduce `AnnounceOnInitMixin` to standardize announcement behavior

This commit introduces a new mixin, `AnnounceOnInitMixin`, designed to provide a consistent way for widgets to make an announcement when they are initialized.

* Refactor MenuScreen to use ListView for improved a11y

Replaced `CustomScrollView` with `ListView` in `MenuScreen` to enhance accessibility support. This change addresses an issue where `CustomScrollView` prevented scrolling to certain elements, such as the 'logout' button, when using accessibility services. Using `ListView` provides a simpler and more reliable solution for this scenario.