Use trait for validating disclosure against DCQL
This introduces the `DisclosedCredential` to the disclosure validation
section of the `dcql` crate. This trait is used to decouple that crate
from the actual disclosure formats and thus removes its dependency on
the `mdoc` crate.

The `DisclosedCredential` is implemented on `DisclosedAttestations`,
which both simplifies the OpenID4VP response verification code and
allows for the DQCL request validation to be format agnostic, as both
mdoc and SD-JWT have their disclosure data converted to this type.