PVW-4589: Recover Pin (UI)
* Add TODOs for removing legacy forgot PIN flow

* Refactor: Rename `CreatePinRecoveryUrlXx` to `CreatePinRecoveryRedirectUriXx`

* Update l10n and associated goldens

* Refactor GetPinRecoveryUrlXx to CreatePinRecoveryUrlXx

Update the method/class names to match the core api.

* Adjust stepper in PIN recovery flow

Hide stepper in loading and error states.

* Add TODOs to handle DigidMismatch

* Fix dart lint issue

* Differentiate PIN recovery/reset flows by build

Mock build supports recovery and thus shows recovery flow
Core build does not yet support recover so still shows the reset flow

* Unlock wallet after completing PIN recovery (mock)

* Update release notes to include pin recovery feature

* Update StopSheet golden (updated l10n)

* Update dutch (nl) l10n

* Fix test name typo

* Add RecoverPinStopSheet golden test

* Add translations for PIN recovery stop sheet

* Add unit tests for RecoverPinEvent

* Update landscape layout in PinSetupPage

Make sure the title renders vertically centered if no onStopPressed is provided

* Refactor PinSetupPage to remove redundant showInput property

This commit refactors the `PinSetupPage` widget by removing the `showInput` property and its associated logic. The PIN keyboard is now always visible.

* Implement RecoverPin BLoC tests

* Add RecoverPinConfirmPinFailed retry golden test

* Refactor PIN confirmation in RecoverPinBloc

- Split out logic of enter & confirm into dedicated methods
- Handle the retry confirmation flow
- Remove unused events

* Localize Recover PIN screen (en)

* Remove stepper from Error & Stopped states

* Add golden tests for RecoverPinScreen

This commit introduces golden tests for the `RecoverPinScreen`, verifying all possible UI states defined in recover_pin_state.

* Update PIN recovery mismatch error page

This commit updates the `RecoverPinMismatch` error page by:
- Adding an error illustration.
- Adding an icon to the secondary button.

* Update error handling in PIN recovery screen

This commit refactors the error handling in the PIN recovery screen:
- Replaces `TerminalPage` with specific `ErrorPage` instances for different error types (`RecoverPinNetworkError`, `RecoverPinGenericError`, `RecoverPinSessionExpired`).
- Implements distinct error pages for network errors with and without internet connectivity.
- Adds a retry mechanism to all error pages, triggering `RecoverPinRetryPressed` event.

* Refine RecoverPinDigidMismatch error screen

* Reduce rebuilds

Avoid rebuilds when they are not needed since the UI does not require an update.

* Handle PIN confirmation failure in PIN recovery flow

This commit introduces the `RecoverPinConfirmPinFailed` state to the `RecoverPinBloc` to handle cases where the entered PIN does not match the chosen PIN during the PIN recovery process.

* Refactor PinSetupPage and add stop button

This commit refactors the `PinSetupPage` to improve its structure and introduces a new "Stop" button.

* Update PIN recovery screen back/stop/pop logic

Consolidate the canPop and canStop logic into dedicated methods, making it easier
to see/alter when the back and close buttons should be shown.

* Update RecoverPinScreen back navigation

Make sure the back icon is visible when the user can pop the recover pin screen, and make it so that
it actually pops the screen in this case.

* Update PIN recovery success page

This commit updates the success page of the PIN recovery flow:
- Add the PIN_SET illustration
- Modifies the primary button to navigate to the dashboard
- Adds a secondary button to navigate to the wallet history

* Implement PIN entry in PIN recovery flow

This commit updates the PIN entry logic within the PIN recovery flow by:
- Migrating to the common `PinSetupPage`.
- Moving the entry logic into the RecoverPinBloc

* Refactor SetupSecurityPinPage into PinSetupPage

This commit moves `SetupSecurityPinPage` to `feature/pin/pin_setup_page.dart` and renames it to `PinSetupPage`.
By doing so the same page can be re-used in the recover pin flow.

* Add placeholder pin entry pages

* Refactor ForgotPinScreen navigation

Use named navigation so we can use popUntil elsewhere.

* Update RecoverPinScreen UI (up to choose pin)

Handle everything up to (and including) the (mock) digid login.

* Handle DigID login user cancellation in BLoC

* Enable PIN recovery deeplink navigation

Introduce the PinRecoveryNavigationRequest (and return it accordingly when identifying associated uri) to enable
deeplink back into the app to the pin recovery page.

* Add Forgot PIN screen route to WalletRoutes

* Add PIN recovery stop confirmation sheet

This commit introduces the `RecoverPinStopSheet` widget. Still needs to be localized correctly.

* Handle user cancellation in DigiD login for PID renewal

This commit updates the `RenewPidBloc` to specifically handle cases where the user cancels the DigiD login process during PID renewal.

* Start building Recover PIN UI pages

This commit sets up the framework to implement the recovery pin pages and starts implementing (hardcoded) variants
of some of the initial pages.

* Pass width and height to SvgOrImage

This commit updates the `SvgOrImage` widget to accept optional `width` and `height` parameters, which are passed through to the child widget.

* Update `RecoverPinScreen` title based on BLoC state

* Update golden files for PIN recovery

* Setup placeholder PIN recovery screen

Create the initial PinRecoveryScreen and provide it with the PinRecoveryBloc.
Also update the ForgotPinScreen cta so it refers to pin recovery and navigates to the recovery screen.

* Reorder route constants alphabetically

This commit reorders the static route constants in `wallet_routes.dart` alphabetically.

* Update RecoverPinBloc to use new PIN recovery use cases

* Refactor PIN recovery use cases

This commit refactors the PIN recovery use cases by renaming them from `Usecase` to `UseCase` for consistency.

* Implement PIN recovery usecases

* Inject PinRepository into PinRecovery Usecases

* Add `PinRepository` and its core implementation

* Provide PIN recovery use cases

* Add PIN recovery use cases

This commit introduces stub implementations for the PIN recovery use cases:
- `GetPinRecoveryUrlUsecaseImpl`
- `ContinuePinRecoveryUsecaseImpl`
- `CompletePinRecoveryUsecaseImpl`
- `CancelPinRecoveryUsecaseImpl`

* Implement mock PIN recovery flow APIs

* Implement RecoverPin BLoC

This commit introduces the `RecoverPinBloc`, along with its corresponding events and states, to manage the PIN recovery process.

* Implement mock PIN recovery flow apis

* Refactor PIN recovery API

This commit refactors the PIN recovery API by:
- Renaming `continue_recover_pin` to `continue_pin_recovery`.
- Splitting the original `continue_recover_pin` into two functions:
    - `continue_pin_recovery`: Takes the recovery URI as input.
    - `complete_pin_recovery`: Takes the new PIN as input.
- Adding a new `cancel_pin_recovery` function.

* Update generated mocks

* Format generated code

* Add API stubs for PIN recovery

This commit introduces new API endpoints and corresponding stubs for the PIN recovery functionality:
- `create_pin_recovery_redirect_uri`: Generates a redirect URI for initiating PIN recovery.
- `continue_recover_pin`: Continues the PIN recovery process using the provided URI and new PIN.

* Add `PinRecovery` URI type

Introduces a new `PinRecovery` enum type in the Rust core and Flutter API layers.