PVW-4602: Wallet Transfer (Source) UI
* Update wallet transfer icon in settings screen

* Wrap WalletTransferFaqScreen body with SafeArea

* Update golden file for issuance stop_sheet

* Update _buildTitle to avoid returning null

Returning null conflicts with an assert to handle the scroll behaviour, so returning an
empty placeholder instead.

* Remove unused translations for PIN recovery

* Use localized titles in WalletTransferSourceScreen

This change updates the `WalletTransferSourceScreen` to use localized titles based on the current state of the `WalletTransferSourceBloc`.

* Implement l10n for WalletTransferSourceStopSheet

* Add "Move NL Wallet" translations

* Add To-Do comments for transfer use cases

Since the core hasn't implemented these features yet, annotate them with TODO to make
sure they are implemented at a later stage.

* Fix typo in comment

* Add tests for WalletTransferSourceBloc

This commit introduces a new test suite for the `WalletTransferSourceBloc`.
The tests cover the following scenarios:
- Initial state verification.
- Emission of `WalletTransferGenericError` when transfer initiation fails.
- Successful transfer flow ("happy path").
- Handling of the `WalletTransferStopRequestedEvent`.
- Handling of the `WalletTransferBackPressedEvent` from the confirm PIN state.
- Handling of a generic error during the transfer process, leading to `WalletTransferFailed`.

* Add golden test for transfer Stop Sheet

* Fix lint issues

* Update l10n and associated goldens

* Add release note for Wallet Transfer (source) flow

* Pass specific ApplicationError from status stream to `_handleError`

* Refactor WalletTransferSourceBloc status handling

This commit refactors the handling of wallet transfer status updates in `WalletTransferSourceBloc`.

- Introduced a `StreamSubscription` to manage the status updates from `_getWalletTransferStatusUseCase`.
- Ensured the subscription is cancelled when a new PIN confirmation event occurs, when a stop is requested, or when the BLoC is closed.
- Modified `_onPinConfirmed` to use the `StreamSubscription` and emit states based on the received `WalletTransferStatus`.

* Handle network and session errors in `WalletTransferSourceBloc`

* Add mocks for Wallet Transfer use cases for testing

* Introduce WalletTransfer related use cases

This commit adds the following use cases for managing wallet transfers:
- `GetWalletTransferStatusUseCase`: Observes the status of an ongoing transfer.
- `InitWalletTransferUseCase`: Initializes a new wallet transfer.
- `StartWalletTransferUseCase`: Starts a wallet transfer after PIN confirmation.
- `CancelWalletTransferUseCase`: Cancels an ongoing wallet transfer.

These use cases are integrated into `WalletTransferSourceBloc` to handle the transfer flow and update the UI accordingly. The `WalletTransferSourceConfirmPinPage` now uses `StartWalletTransferUseCase` to initiate the transfer after successful PIN entry.

* Reset some english translations

Incorrectly imported some translations from a different (recover pin) MR which should not yet
be part of this branch, reverting as such.

* Update dutch transfer translations

* Internationalize wallet transfer source flow (l10n)

This commit replaces hardcoded strings in the wallet transfer source user flow with localized keys from the `.arb` files.

* Localize WalletTransferFaqScreen

* Make WalletTransfer naming more explicit

Transfer -> WalletTransfer

* Add transfer wallet FAQ screen

This introduces a new FAQ screen that explains how to move the NL Wallet to a new phone.

* Fix lint issues

* Refactor navigation requests using factories

- Introduce factory constructors for different `NavigationRequest` types to simplify their creation.
- Create a shared `unlockedWithPidPrerequisites` constant to reduce code duplication for common navigation checks.
- Update the test `text_styles_tab` to use the new `NavigationRequest.transfer` factory.

* Handle pop behaviour

- Allow users to pop the screen terminal (error) states
- Trigger stop sheet on pop attempts on non-terminal states

* Extract stop confirmation logic into `WalletTransferSourceStopSheet`

This change introduces the `WalletTransferSourceStopSheet` to encapsulate the UI for confirming the cancellation of a wallet transfer.

* Add back support to WalletTransfer

* Disable automaticallyImplyLeading during wallet transfer (source)

* Render WalletTranferring state

- Show the WalletTransferSourceTransferringPage() when transferring
- Add a ScrollOffSetProvider to manage the scroll offset
- Update tranferring goldens

* Create WalletTransferSourceTransferringPage

This commit introduces the `WalletTransferSourceTransferringPage`, a new page that is to be displayed while the wallet is being transferred.

* Add wallet transfer (source) dummy entry

- Allow entry to the wallet transfer mock from the design page
- Update bloc to support mock flow

* Add golden test for transfer failed details sheet

* Implement WalletTransfer 'failure' UI state

* Implement common error pages for WalletTransfer

Generic/Network/Session errors now refer to their common error page implementation.

* Implement `WalletTransferStopped` page

This change implements the user interface for the `WalletTransferStopped` state.

* Implement transfer success page for source device

- Add a new page `WalletTransferSourceTransferSuccessPage` to be displayed on the source device after a successful wallet transfer.
- Include a new illustration (`NL_WALLET_move_source_success.svg`) for the success page.
- Add a "Close" button for Android users to exit the application.

* Add loading animation for WalletTransferTransferring state

This commit introduces a Lottie animation to the WalletTransferTransferring state of the wallet transfer process.

* Add WalletTransferSourceConfirmPinPage

This commit introduces the `WalletTransferSourceConfirmPinPage` widget, which is displayed when the user needs to confirm their PIN during a wallet transfer.

* Add progress indicator to wallet transfer

This commit introduces a progress indicator to the wallet transfer flow.

* Add transfer source introduction page

This commit introduces the initial version of the wallet transfer source introduction page.

* Add golden tests for WalletTransferSourceScreen

This commit introduces golden tests for the `WalletTransferSourceScreen` and all it's possible states.
This is done before the implementation of the actual UI states as a means of test driven development, all states
are rendered in light mode and in dark+landscape mode to cover the most common usecases.

* Implement placeholder UI states

* Add `canGoBack` property to `WalletTransferSourceState`

* Fix mis-named InitiateTransferEvent call

* Implement basics of WalletTransferSourceBloc

* Add `WalletTransferStopped` state

* Implement the required WalletTransferSourceEvents

* Expand WalletTransferSource states

This commit introduces a more comprehensive set of states to the `WalletTransferSourceBloc` to represent the various stages and potential outcomes of a wallet transfer process.

* Add WalletTransferSource feature skeleton

This commit introduces the `WalletTransferSource` feature, which includes the following components:

- `WalletTransferSourceBloc`: Manages the state for the wallet transfer source flow
- `WalletTransferSourceEvent`: Defines events related to wallet transfer
- `WalletTransferSourceState`: Represents the different states of the wallet transfer process
- `WalletTransferSourceScreen`: Provides the UI for the wallet transfer source feature.