PVW-4911: [A11y] Make CTAs non-sticky
* Fix lint issue

* Re-introduce Scrollbar in TerminalPage & update goldens

* Refactor widget structure in IntroductionPrivacyScreen

This commit removes an unnecessary `Expanded` widget from the `_buildContent` method in the `IntroductionPrivacyScreen`.

* Fix TerminalPage layout to make button un-sticky

Update the TerminalPage so that the buttons are not sticky at the bottom of the screen, but instead
are part of the scroll view if the scroll content would not fit naturally.

* Make IntroductionPrivacyScreen bottom section non-sticky

Move the buttons so that they are part of the scrollview, making sure the viewport is as big
as possible on smaller devices with significant font scaling enabled.