PVW-4924: Introduce PinRecoveryMethod for ForgotPinScreen
* Refactor l10n keys for forgot PIN screen

* Introduce PinRecoveryMethod for ForgotPinScreen

This commit introduces a `PinRecoveryMethod` enum to allow the `ForgotPinScreen` to support different user flows: PIN recovery and wallet reset.
This is then used during the last step of the personalization flow (add Pid) to fallback to the reset wallet flow, since at that stage there is
no recovery code yet and thus the only actionable approach is to reset the wallet and start over.