Remove non-requested disclosed attributes
Have the verifier remove any received attributes that were not present
in the DCQL credential request it sent to the holder. This filters out
both metadata present in the SD-JWT payload and any non-selectively
disclosable attributes that were not requested.

For the sake of consistency this is also done for mdoc, just in case the
holder sends any attributes that were not requested for some reason.