PVW-4911: Tweak PinLayout for a11y
* Regenerate golden file for PIN page

* Refactor `reduceSpacing` for more generic support

The 'mediaQuery.size.height < 600' caused all iOS devices to be excluded, we now also opt
to reduce the spacings when the textScaling is set to really extreme values to make sure the
ui behaves as expected in these situations as well.

* Update pin-related goldens

* Refactor PinPage layout for landscape and adjust padding

* Improve UI on small devices with large text scaling

This commit introduces adjustments to the UI to better accommodate small devices, especially when large text scaling is enabled for accessibility.

* Refactor: Adjust PinKeyboard height constraints

This commit reduces the maximum height of the `PinKeyboard` widget.

- Decreased `_maxHeight` from 340.0 to 328.0.
- Reduced `_maxHeightAsFractionOfScreen` from 0.44 to 0.35 to improve layout on smaller screens.

* Remove "je" from Dutch translations for pin confirmation