PVW-4878: Connect transfer core apis
* Update copy on wallet transfer confirmation screen

* Refactor WalletTransferStatus enum and add docs

* Implement SkipWalletTransferUseCase

This commit implements the `SkipWalletTransferUseCase` to allow skipping the wallet transfer process.

* Add basic test for SemanticsEventService

* Update GetWalletTransferStatusUseCase and add tests

- Removed the `isTarget` parameter from `GetWalletTransferStatusUseCase.invoke` as it was only used by mocks.
- Added `MockTransferRepository` to `wallet_mocks.dart`.
- Added unit tests for `GetWalletTransferStatusUseCaseImpl`.

* Update PID repository tests and add new test cases

* Add tests for QR code navigation requests

This commit adds tests for the following QR code types:
- Pid Renewal
- Pin Recovery
- Disclosure Based Issuance
- Transfer

* Refactor NavigationRequest and update tests

This commit refactors the `NavigationRequest` factory constructors and updates the corresponding tests in `CoreUriRepositoryTest`.

* Add unit tests for CoreTransferRepository

* Support wallet transfer navigation via QR code and URI

* Mark `_eventLog` as unused

* Implement wallet transfer mock logic

This commit moves `PinManager`, `IssuanceManager`, and `DisclosureManager` into a new `manager` directory. It also introduces `TransferManager` to handle the mock logic for wallet transfer functionality.

* Implement StartWalletTransferUseCase

This commit implements the `StartWalletTransferUseCase` to initiate a wallet transfer process.

* Refactor PrepareWalletTransferUseCase to InitWalletTransferUseCase

This commit renames `PrepareWalletTransferUseCase` to `InitWalletTransferUseCase` and updates its implementation to call the `initWalletTransfer` method from the `TransferRepository`.

* Refactor WalletTransferSourceBloc tests to use AcknowledgeWalletTransferUseCase

* Rename InitWalletTransferUseCase to AcknowledgeWalletTransferUseCase

This commit renames `InitWalletTransferUseCase` to `AcknowledgeWalletTransferUseCase` and implements it by calling through to the wallet_core.

* Implement CancelWalletTransferUseCase

* Introduce TransferRepository and integrate with Wallet Core

This commit introduces a `TransferRepository` to manage wallet transfer operations and integrates it with the Wallet Core. The get_wallet_transfer_status_usecase is also update to use a polling mechanism
to get the latest transfer status.

* Pop TransferTargetScreen

To make sure the user ends up in the correct spot after skipping the transfer we now
simply pop the screen, which should bring the user back to the end of the personalize flow.

* Redirect to transfer screen after personalization if available

This commit updates the wallet personalization flow to redirect the user to the wallet transfer screen if the transfer option is available after successfully adding the PID.

* Fix test after core API rename

* Pass TransferState on PIN confirmation

This commit updates the `WalletPersonalizePinConfirmed` event to include the `transferState`. This allows the `WalletPersonalizeBloc` to access the result from the PIN confirmation step.

* Refactor PID acceptance flow and introduce TransferState

Refactor pid acceptance to use dedicated acceptPid api and to bubble up the TransferState