PVW-4784: Need Help & Contact UI
* Regenerate pin goldens for copy change

(pin->PIN)

* Sync translations

* Update security icon on Need Help screen

* Update release notes to include need help & contact placeholder pages

* Update menu related golden tests for landscape mode

This commit updates the golden file tests for `MenuScreen` and `SettingsScreen` to use landscape orientation for the dark mode tests.

* Add golden tests for ContactScreen

This commit introduces golden file tests for the `ContactScreen` to verify its appearance.

- Added tests for both light and dark modes.
- Included a test case for landscape orientation.

* Implement UI of Contact screen

* Make BulletListDot the default icon for BulletList

* Replace placeholder with navigation to ContactScreen

This commit wires up the "Contact" button in the "Need Help" screen to navigate to the new `ContactScreen`.

* Add initial ContactScreen

* Fix icon color on NeedHelpScreen and add golden tests

* Replace share icon with disclose icon on Need Help screen

* Update Help menu item to navigate to NeedHelpScreen

* Add Need Help screen route

This commit introduces a new route for the "Need Help" screen.

- Added `needHelpRoute` constant and its corresponding builder in `wallet_routes.dart`.
- Updated the `settingsRoute` path to also be nested under `/menu`.

* Introduce "Need Help?" screen

This commit adds a new "Need Help?" screen, accessible from the main menu. The screen provides a list of help topics, which currently lead to placeholder pages.

* Move SettingsScreen into menu feature

This commit moves the `SettingsScreen` and its related files from the generic `feature/settings` directory to `feature/menu/sub_menu/settings`. This refactoring places the screen within the menu feature, which is its logical parent.

* Pop to dashboard when app is locked

This commit updates the `MenuScreen` and `SettingsScreen` to navigate back to the dashboard when the app gets locked.