PVW-5048: Transfer UI Tweaks
* Update cancel button icon on WalletTransferAwaitingConfirmationPage

* Remove close button from transferring screen

This commit removes the close (`X`) button from the app bar on the wallet transfer screens for both the source and target devices during the "transferring" state.

* Update stop icon on WalletTransferringPage

This commit changes the icon used for the "Stop" button on the `WalletTransferringPage` from `Icons.close_outlined` to `Icons.block_flipped` for better visual representation of the action.